Rework Client type to match SQL; separate Client from Request in ClientDialog state
